+++ client/src/components/Bookshelf.js
@@ -1,0 +1,11 @@
+import React, { Component } from 'react';
+
+class BookShelf extends Component {
+  render() {
+    return (
+      <div>BookShelf!</div>
+    );
+  }
+}
+
+export default BookShelf;
