+++ client/src/components/Header.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import logo from '../logo.svg';
+
+const Header = () => {
+  return (
+    <header className="App-header">
+      <img src={logo} className="App-logo" alt="logo" />
+      <h2>Sharing BookShelf</h2>
+    </header>
+  );
+}
+
+export default Header;
